added a plex setup script

+++ setup.sh
@@ -59,12 +59,14 @@ chown ${USERNAME}:${GROUPNAME} ${DDCLIENT_CONF_DIR}/ddclient.conf
 chmod 640 ${DDCLIENT_CONF_DIR}/ddclient.conf
 
 # Create acme.json file for letsencrypt
+if [[ ! -f ${CONFIGS_BASE_DIR}/letsencrypt/acme.json ]]
+then
+  [[ -f ./acme.json ]] && cp ./acme.json ${CONFIGS_BASE_DIR}/letsencrypt/acme.json
+fi
 mkdir -p ${CONFIGS_BASE_DIR}/letsencrypt
 touch ${CONFIGS_BASE_DIR}/letsencrypt/acme.json
 chmod 600 ${CONFIGS_BASE_DIR}/letsencrypt/acme.json
 chown ${USERNAME}:${GROUPNAME} ${CONFIGS_BASE_DIR}/letsencrypt/acme.json
-[[ -f ./acme.json ]] && cp ./acme.json ${CONFIGS_BASE_DIR}/letsencrypt/acme.json
-
 
 # Create docker-compose.override.yml
 if [[ ! -f docker-compose.override.yml ]]
